test(tui): consolidate picker rail coverage
